Organizing the code +
Refactoring the project's structure to follow the Model-View-Controller (MVC) pattern as recommended by our professors. In addition, the quizzes have been added. While there are still some coding connections to be made, the project structure has improved significantly and is coming together nicely. Progressing with the genius touch as always.